LPS-148908 Moves PropsTemporarySwapper to Props utility module to be able to use it across the test modules.
