+++ modules/apps/portal/portal-props-test-util/src/main/resources/com/liferay/portal/props/test/util/packageinfo
@@ -1,0 +1,1 @@
+version 2.2.0
